+++ tests/test_mesh.py
@@ -44,6 +44,11 @@ def test_meshes():
         cell_type="triangle",
     )
 
+    n = m.copy()
+    n.update(m.cells, cell_type="my-fancy-cell-type")
+
+    assert n.cell_type == "my-fancy-cell-type"
+
     fe.mesh.convert(m, order=0)
     fe.mesh.convert(m, order=0, calc_points=True)
     fe.mesh.convert(m, order=2)